FE/Qt4: fix warning
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@21437 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Frontends/VirtualBox/src/QIAdvancedSlider.cpp
@@ -268,7 +268,7 @@ int QIAdvancedSlider::snapValue(int val)
         val > 2)
     {
         float l2 = log ((float)val)/log (2.0);
-        int newVal = pow ((float)2, (int)qRound (l2)); /* The value to snap on */
+        int newVal = (int) pow ((float)2, (int)qRound (l2)); /* The value to snap on */
         int pos = mSlider->positionForValue (val); /* Get the relative screen pos for the original value */
         int newPos = mSlider->positionForValue (newVal); /* Get the relative screen pos for the snap value */
         if (abs (newPos - pos) < 5) /* 10 pixel snapping range */
